+++ src/ThreeDAnalyzer.Web/obj/Debug/net10.0/ThreeDAnalyzer.Web.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+64bb7de16ba79eaf46743107a6d8310a560ca7bb")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+6f0846f709c19cbd8f1d5039518bca196c690d0b")]
 [assembly: System.Reflection.AssemblyProductAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyTitleAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]